src/main/java/ru/skypro/questions/service/impl/JavaQuestionService.java
package ru.skypro.questions.service.impl;

import jakarta.annotation.PostConstruct;
import org.springframework.stereotype.Service;
import ru.skypro.questions.model.Question;
import ru.skypro.questions.service.QuestionService;

import java.util.*;

@Service

public class JavaQuestionService implements QuestionService {
    private final Set<Question> questions = new HashSet<>();
    private final Random random = new Random();
    @PostConstruct
    public void init() {
        add("переменная какого типа возвращает true/false?", "переменная boolean");
        add("является ли Map коллекцией?", "не является");
        add("где хранятся примитивные переменные?", "в Стеке");

    }
    @Override
    public Question add(String question, String answer) {
        Question newQuestion = new Question(question, answer);
        questions.add(newQuestion);
        return newQuestion;
    }

    @Override
    public Question add(Question question) {
        questions.add(question);
        return question;
    }

    @Override
    public Question remove(Question question) {
        questions.remove(question);
        return question;
    }

    @Override
    public Collection<Question> getAll() {
        return Collections.unmodifiableSet(questions);
    }

    @Override
    public Question getRandomQuestion() {
        int randomIndex = random.nextInt(questions.size());
        return new ArrayList<>(questions).get(randomIndex);
    }
}
